+++ flang/lib/Optimizer/Transforms/MIFOpConversion.cpp
@@ -430,7 +430,7 @@ struct MIFThisImageOpConversion
     mlir::Location loc = op.getLoc();
 
     if (op.getCoarray())
-      TODO(loc, "mif.this_image op with coarray argument.");
+      TODO(loc, "coarray: mif.this_image op with coarray argument");
     else {
       mlir::Type i32Ty = builder.getI32Type();
       mlir::Type boxTy = fir::BoxType::get(rewriter.getNoneType());